+++ libmariadb/mariadb_lib.c
@@ -1944,7 +1944,7 @@ mysql_query(MYSQL *mysql, const char *query)
 */
 
 int STDCALL
-mysql_send_query(MYSQL* mysql, const char* query, size_t length)
+mysql_send_query(MYSQL* mysql, const char* query, unsigned long length)
 {
   return ma_simple_command(mysql, COM_QUERY, query, length, 1,0);
 }
@@ -2141,11 +2141,11 @@ mysql_read_query_result(MYSQL *mysql)
 }
 
 int STDCALL
-mysql_real_query(MYSQL *mysql, const char *query, size_t length)
+mysql_real_query(MYSQL *mysql, const char *query, unsigned long length)
 {
   my_bool skip_result= OPT_EXT_VAL(mysql, multi_command);
 
-  if (length == (size_t)-1)
+  if (length == (unsigned long)-1)
     length= strlen(query);
 
   free_old_query(mysql);
